Switched to multi-root workspaces
Moved Java build logic from the root folder to the ./VM folder. VSCode does not like it otherwise anymore. Made equivalent VSCode multi-root workspaces for the root, ./VM and ./Smalltalk folders for convenient startup.
The root folder itself may not be part of a workspace anymore, because VSCode gets confused about what to build and there is no way to specify a startup project.